# src/routes.ts
/* tslint:disable */
/* eslint-disable */
// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
import {Controller, ValidationService, FieldErrors, ValidateError, TsoaRoute} from 'tsoa';
// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
import {UserController} from './controllers/UserController';
// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
import {AuthenticationController} from './controllers/AuthenticationController';
import {expressAuthentication} from './middlewares/Authenticator';
import * as express from 'express';

// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

const models: TsoaRoute.Models = {
    "IUser": {
        "dataType": "refObject",
        "properties": {
            "username": {"dataType": "string", "required": true},
            "roles": {"dataType": "array", "array": {"dataType": "string"}, "required": true},
            "email": {"dataType": "string", "required": true},
            "name": {"dataType": "string", "required": true},
            "active": {"dataType": "enum", "enums": ["true"], "required": true},
        },
        "additionalProperties": true,
    },
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    "IPaginateResultIUser": {
        "dataType": "refObject",
        "properties": {
            "docs": {"dataType": "array", "array": {"ref": "IUser"}, "required": true},
            "total": {"dataType": "double", "required": true},
            "limit": {"dataType": "double", "required": true},
            "page": {"dataType": "double"},
            "pages": {"dataType": "double"},
            "offset": {"dataType": "double"},
        },
        "additionalProperties": true,
    },
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    "ISearchAndFilter": {
        "dataType": "refObject",
        "properties": {
            "query": {"dataType": "any"},
            "filter": {"dataType": "any"},
        },
        "additionalProperties": true,
    },
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    "ICreateUserDto": {
        "dataType": "refObject",
        "properties": {
            "username": {"dataType": "string", "required": true},
            "email": {"dataType": "string", "required": true},
            "name": {"dataType": "string", "required": true},
            "password": {"dataType": "string", "required": true},
        },
        "additionalProperties": true,
    },
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    "IAuthenticationResponse": {
        "dataType": "refObject",
        "properties": {
            "token": {"dataType": "string", "required": true},
            "user": {"ref": "IUser", "required": true},
        },
        "additionalProperties": true,
    },
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    "ILogin": {
        "dataType": "refObject",
        "properties": {
            "username": {"dataType": "string", "required": true},
            "password": {"dataType": "string", "required": true},
        },
        "additionalProperties": true,
    },
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
};
const validationService = new ValidationService(models);

// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

export function RegisterRoutes(app: express.Express) {
    // ###########################################################################################################
    //  NOTE: If you do not see routes for all of your controllers in this file, then you might not have informed tsoa of where to look
    //      Please look into the "controllerPathGlobs" config option described in the readme: https://github.com/lukeautry/tsoa
    // ###########################################################################################################
    app.get('/api/v1/user/me',
        authenticateMiddleware([{"jwt": []}]),
        function (request: any, response: any, next: any) {
            const args = {
                request: {"in": "request", "name": "request", "required": true, "dataType": "object"},
            };

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            let validatedArgs: any[] = [];
            try {
                validatedArgs = getValidatedArgs(args, request);
            } catch (err) {
                return next(err);
            }

            const controller = new UserController();


            const promise = controller.getMe.apply(controller, validatedArgs as any);
            promiseHandler(controller, promise, response, next);
        });
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    app.get('/api/v1/user',
        authenticateMiddleware([{"jwt": ["admin"]}]),
        function (request: any, response: any, next: any) {
            const args = {
                page: {"default": 1, "in": "query", "name": "page", "dataType": "double"},
                limit: {"default": 10, "in": "query", "name": "limit", "dataType": "double"},
                sortAsc: {"default": true, "in": "query", "name": "sortAsc", "dataType": "boolean"},
                fieldSort: {"default": "_id", "in": "query", "name": "fieldSort", "dataType": "string"},
            };

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            let validatedArgs: any[] = [];
            try {
                validatedArgs = getValidatedArgs(args, request);
            } catch (err) {
                return next(err);
            }

            const controller = new UserController();


            const promise = controller.getAll.apply(controller, validatedArgs as any);
            promiseHandler(controller, promise, response, next);
        });
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    app.post('/api/v1/user/filter',
        authenticateMiddleware([{"jwt": ["admin"]}]),
        function (request: any, response: any, next: any) {
            const args = {
                body: {"in": "body", "name": "body", "required": true, "ref": "ISearchAndFilter"},
                page: {"default": 1, "in": "query", "name": "page", "dataType": "double"},
                limit: {"default": 10, "in": "query", "name": "limit", "dataType": "double"},
                sortAsc: {"default": true, "in": "query", "name": "sortAsc", "dataType": "boolean"},
                fieldSort: {"default": "_id", "in": "query", "name": "fieldSort", "dataType": "string"},
            };

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            let validatedArgs: any[] = [];
            try {
                validatedArgs = getValidatedArgs(args, request);
            } catch (err) {
                return next(err);
            }

            const controller = new UserController();


            const promise = controller.filter.apply(controller, validatedArgs as any);
            promiseHandler(controller, promise, response, next);
        });
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    app.get('/api/v1/user/:id',
        authenticateMiddleware([{"jwt": ["admin"]}]),
        function (request: any, response: any, next: any) {
            const args = {
                id: {"in": "path", "name": "id", "required": true, "dataType": "string"},
                request: {"in": "request", "name": "request", "required": true, "dataType": "object"},
            };

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            let validatedArgs: any[] = [];
            try {
                validatedArgs = getValidatedArgs(args, request);
            } catch (err) {
                return next(err);
            }

            const controller = new UserController();


            const promise = controller.get.apply(controller, validatedArgs as any);
            promiseHandler(controller, promise, response, next);
        });
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    app.post('/api/v1/user',
        authenticateMiddleware([{"jwt": ["admin"]}]),
        function (request: any, response: any, next: any) {
            const args = {
                body: {"in": "body", "name": "body", "required": true, "ref": "ICreateUserDto"},
                request: {"in": "request", "name": "request", "required": true, "dataType": "object"},
            };

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            let validatedArgs: any[] = [];
            try {
                validatedArgs = getValidatedArgs(args, request);
            } catch (err) {
                return next(err);
            }

            const controller = new UserController();


            const promise = controller.create.apply(controller, validatedArgs as any);
            promiseHandler(controller, promise, response, next);
        });
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    app.put('/api/v1/user/:id',
        authenticateMiddleware([{"jwt": ["admin"]}]),
        function (request: any, response: any, next: any) {
            const args = {
                id: {"in": "path", "name": "id", "required": true, "dataType": "string"},
                body: {"in": "body", "name": "body", "required": true, "ref": "IUser"},
                request: {"in": "request", "name": "request", "required": true, "dataType": "object"},
            };

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            let validatedArgs: any[] = [];
            try {
                validatedArgs = getValidatedArgs(args, request);
            } catch (err) {
                return next(err);
            }

            const controller = new UserController();


            const promise = controller.update.apply(controller, validatedArgs as any);
            promiseHandler(controller, promise, response, next);
        });
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    app.delete('/api/v1/user/:id',
        authenticateMiddleware([{"jwt": ["admin"]}]),
        function (request: any, response: any, next: any) {
            const args = {
                id: {"in": "path", "name": "id", "required": true, "dataType": "string"},
                body: {"in": "body", "name": "body", "required": true, "ref": "IUser"},
                request: {"in": "request", "name": "request", "required": true, "dataType": "object"},
            };

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            let validatedArgs: any[] = [];
            try {
                validatedArgs = getValidatedArgs(args, request);
            } catch (err) {
                return next(err);
            }

            const controller = new UserController();


            const promise = controller.delete.apply(controller, validatedArgs as any);
            promiseHandler(controller, promise, response, next);
        });
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    app.post('/api/v1/authentication/signup',
        function (request: any, response: any, next: any) {
            const args = {
                body: {"in": "body", "name": "body", "required": true, "ref": "ICreateUserDto"},
                request: {"in": "request", "name": "request", "required": true, "dataType": "object"},
            };

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            let validatedArgs: any[] = [];
            try {
                validatedArgs = getValidatedArgs(args, request);
            } catch (err) {
                return next(err);
            }

            const controller = new AuthenticationController();


            const promise = controller.signup.apply(controller, validatedArgs as any);
            promiseHandler(controller, promise, response, next);
        });
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
    app.post('/api/v1/authentication/login',
        function (request: any, response: any, next: any) {
            const args = {
                body: {"in": "body", "name": "body", "required": true, "ref": "ILogin"},
                request: {"in": "request", "name": "request", "required": true, "dataType": "object"},
            };

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            let validatedArgs: any[] = [];
            try {
                validatedArgs = getValidatedArgs(args, request);
            } catch (err) {
                return next(err);
            }

            const controller = new AuthenticationController();


            const promise = controller.login.apply(controller, validatedArgs as any);
            promiseHandler(controller, promise, response, next);
        });
    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

    function authenticateMiddleware(security: TsoaRoute.Security[] = []) {
        return (request: any, _response: any, next: any) => {
            let responded = 0;
            let success = false;

            const succeed = function (user: any) {
                if (!success) {
                    success = true;
                    responded++;
                    request['user'] = user;
                    next();
                }
            }

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            const fail = function (error: any) {
                responded++;
                if (responded == security.length && !success) {
                    error.status = error.status || 401;
                    next(error)
                }
            }

            // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

            for (const secMethod of security) {
                if (Object.keys(secMethod).length > 1) {
                    let promises: Promise<any>[] = [];

                    for (const name in secMethod) {
                        promises.push(expressAuthentication(request, name, secMethod[name]));
                    }

                    Promise.all(promises)
                        .then((users) => {
                            succeed(users[0]);
                        })
                        .catch(fail);
                } else {
                    for (const name in secMethod) {
                        expressAuthentication(request, name, secMethod[name])
                            .then(succeed)
                            .catch(fail);
                    }
                }
            }
        }
    }

    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

    function isController(object: any): object is Controller {
        return 'getHeaders' in object && 'getStatus' in object && 'setStatus' in object;
    }

    function promiseHandler(controllerObj: any, promise: any, response: any, next: any) {
        return Promise.resolve(promise)
            .then((data: any) => {
                let statusCode;
                if (isController(controllerObj)) {
                    const headers = controllerObj.getHeaders();
                    Object.keys(headers).forEach((name: string) => {
                        response.set(name, headers[name]);
                    });

                    statusCode = controllerObj.getStatus();
                }

                // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

                if (data || data === false) { // === false allows boolean result
                    response.status(statusCode || 200).json(data);
                } else {
                    response.status(statusCode || 204).end();
                }
            })
            .catch((error: any) => next(error));
    }

    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa

    function getValidatedArgs(args: any, request: any): any[] {
        const fieldErrors: FieldErrors = {};
        const values = Object.keys(args).map((key) => {
            const name = args[key].name;
            switch (args[key].in) {
                case 'request':
                    return request;
                case 'query':
                    return validationService.ValidateParam(args[key], request.query[name], name, fieldErrors, undefined, {
                        "controllerPathGlobs": ["./src/controllers/UserController.ts", "./src/controllers/AuthenticationController.ts"],
                        "specVersion": 2
                    });
                case 'path':
                    return validationService.ValidateParam(args[key], request.params[name], name, fieldErrors, undefined, {
                        "controllerPathGlobs": ["./src/controllers/UserController.ts", "./src/controllers/AuthenticationController.ts"],
                        "specVersion": 2
                    });
                case 'header':
                    return validationService.ValidateParam(args[key], request.header(name), name, fieldErrors, undefined, {
                        "controllerPathGlobs": ["./src/controllers/UserController.ts", "./src/controllers/AuthenticationController.ts"],
                        "specVersion": 2
                    });
                case 'body':
                    return validationService.ValidateParam(args[key], request.body, name, fieldErrors, name + '.', {
                        "controllerPathGlobs": ["./src/controllers/UserController.ts", "./src/controllers/AuthenticationController.ts"],
                        "specVersion": 2
                    });
                case 'body-prop':
                    return validationService.ValidateParam(args[key], request.body[name], name, fieldErrors, 'body.', {
                        "controllerPathGlobs": ["./src/controllers/UserController.ts", "./src/controllers/AuthenticationController.ts"],
                        "specVersion": 2
                    });
            }
        });

        if (Object.keys(fieldErrors).length > 0) {
            throw new ValidateError(fieldErrors, '');
        }
        return values;
    }

    // WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa
}

// WARNING: This file was auto-generated with tsoa. Please do not modify it. Re-run tsoa to re-generate this file: https://github.com/lukeautry/tsoa